home-work-6/src/main/java/ru/otus/spring/repositories/JpaCommentRepository.java
package ru.otus.spring.repositories;

import jakarta.persistence.EntityManager;
import jakarta.persistence.PersistenceContext;
import jakarta.persistence.TypedQuery;
import org.springframework.stereotype.Repository;
import ru.otus.spring.models.Comment;

import java.util.List;
import java.util.Optional;

@Repository
public class JpaCommentRepository implements CommentRepository {

    @PersistenceContext
    private EntityManager em;

    @Override
    public Optional<Comment> findById(long commentId) {
        TypedQuery<Comment> query = em.createQuery("select c from Comment as c join fetch c.book where c.id = :id",
                Comment.class);
        query.setParameter("id", commentId);
        return Optional.of(query.getSingleResult());
    }

    @Override
    public List<Comment> findByBookId(long bookId) {
        TypedQuery<Comment> query = em.createQuery("select c from Comment as c join fetch c.book where " +
                "c.book.id = :bookId", Comment.class);
        query.setParameter("bookId", bookId);
        return query.getResultList();
    }

    @Override
    public Comment save(Comment comment) {
        if (comment.getId() == null) {
            em.persist(comment);
            return comment;
        } else {
            return em.merge(comment);
        }
    }

    @Override
    public void delete(Comment comment) {
        em.remove(comment);
    }
}
